chore: :truck: internalize dep status bar
