--- packages/status-bar/CHANGELOG.md
@@ -1,60 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## Version [0.0.15]
-
-Replace 0 with -
-
-## Version [0.0.14]
-
-Update formatting styles for numbers
-
-## [0.0.13](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-status-bar@0.0.12...@equinor/workspace-status-bar@0.0.13) (2023-03-27)
-
-**Note:** Version bump only for package @equinor/workspace-status-bar
-
-## [0.0.12](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-status-bar@0.0.11...@equinor/workspace-status-bar@0.0.12) (2023-03-22)
-
-**Note:** Version bump only for package @equinor/workspace-status-bar
-
-## [0.0.11](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-status-bar@0.0.10...@equinor/workspace-status-bar@0.0.11) (2023-02-06)
-
-**Note:** Version bump only for package @equinor/workspace-status-bar
-
-## [0.0.10](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-status-bar@0.0.9...@equinor/workspace-status-bar@0.0.10) (2022-12-07)
-
-**Note:** Version bump only for package @equinor/workspace-status-bar
-
-## [0.0.9](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-status-bar@0.0.8...@equinor/workspace-status-bar@0.0.9) (2022-11-30)
-
-**Note:** Version bump only for package @equinor/workspace-status-bar
-
-## [0.0.8](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-status-bar@0.0.7...@equinor/workspace-status-bar@0.0.8) (2022-11-29)
-
-**Note:** Version bump only for package @equinor/workspace-status-bar
-
-## [0.0.7](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-status-bar@0.0.6...@equinor/workspace-status-bar@0.0.7) (2022-11-10)
-
-**Note:** Version bump only for package @equinor/workspace-status-bar
-
-## [0.0.6](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-status-bar@0.0.4...@equinor/workspace-status-bar@0.0.6) (2022-11-08)
-
-**Note:** Version bump only for package @equinor/workspace-status-bar
-
-## 0.0.5 (2022-11-08)
-
-**Note:** Version bump only for package @equinor/workspace-status-bar
-
-## [0.0.4](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-status-bar@0.0.3...@equinor/workspace-status-bar@0.0.4) (2022-11-08)
-
-**Note:** Version bump only for package @equinor/workspace-status-bar
-
-## [0.0.3](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-status-bar@0.0.2...@equinor/workspace-status-bar@0.0.3) (2022-11-07)
-
-**Note:** Version bump only for package @equinor/workspace-status-bar
-
-## 0.0.2 (2022-11-07)
-
-**Note:** Version bump only for package @equinor/workspace-status-bar
