--- packages/status-bar/README.md
@@ -1,7 +1,0 @@
-# status-bar
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test status-bar` to execute the unit tests via [Jest](https://jestjs.io).
