--- packages/status-bar/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "@equinor/workspace-status-bar",
-  "version": "0.0.15",
-  "type": "module",
-  "sideEffects": false,
-  "license": "MIT",
-  "scripts": {
-    "build": "tsc -b",
-    "build:prod": "tsc -b",
-    "lint": "eslint --ext .ts,.tsx"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "module": "dist/esm/index.js",
-  "types": "dist/types/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "peerDependencies": {
-    "react": ">= 16.8.0",
-    "react-dom": ">= 16.8.0",
-    "react-is": ">= 16.8.0",
-    "styled-components": "^5.3.6"
-  },
-  "dependencies": {
-    "@equinor/eds-tokens": "^0.9.0"
-  },
-  "devDependencies": {
-    "@types/react": ">= 16.8.0",
-    "@types/react-dom": "^18.0.8",
-    "@types/styled-components": "^5.1.26",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-is": "^18.2.0",
-    "styled-components": "^5.3.6",
-    "typescript": "^5.0.2"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/equinor/fusion-workspace.git",
-    "directory": "packages/status-bar"
-  },
-  "gitHead": "6407f12589214b96228ab87d32a211f7c1cd6ba4"
-}
