--- packages/status-bar/src/lib/components/index.ts
@@ -1,1 +1,0 @@
-export * from './StatusBar';
