--- packages/status-bar/src/lib/types/statusItem.ts
@@ -1,7 +1,0 @@
-export interface StatusItem {
-  /** Title to be shown above the value */
-  title: string;
-  /** Value to be shown in the status bar */
-  value: string | number;
-  description?: string;
-}
