+++ packages/workspace-fusion/package.json
@@ -80,7 +80,6 @@
     "@equinor/workspace-powerbi": "workspace:^",
     "@equinor/workspace-react": "workspace:^",
     "@equinor/workspace-sidesheet": "workspace:^",
-    "@equinor/workspace-status-bar": "workspace:^",
     "@tanstack/react-query": "^4.23.0",
     "@types/react-dom": "^18.0.8",
     "history": "^5.3.0",
