+++ packages/workspace-fusion/src/lib/integrations/status-bar/components/statusBar.styles.ts
