+++ packages/workspace-fusion/src/lib/integrations/status-bar/components/statusBarItem.styles.ts
