+++ packages/workspace-fusion/src/lib/integrations/status-bar/types/statusItem.ts
@@ -1,0 +1,7 @@
+export interface StatusItem {
+  /** Title to be shown above the value */
+  title: string;
+  /** Value to be shown in the status bar */
+  value: string | number;
+  description?: string;
+}
